+++ frontend-v2/src/components/client/UserDropdown.tsx
@@ -1,8 +1,8 @@
 "use client"
 
 import { User } from "lucide-react";
-import Link from "next/link";
 
+import { Link } from "@/i18n/routing";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,